core update
commit1:
https://github.com/TrinityCore/TrinityCore/commit/62adbdab992ec177f60dec76447508741de4e7db#

https://github.com/TrinityCore/TrinityCore/commit/9c4cf9b39c17854fac3c1323e5e3bd1d2f164e3d#diff-f7bd8139463608ec43535300fae1ee00bde1ee756f901784729156edf8bb2fed

+++ src/server/game/World/World.h
@@ -849,7 +849,7 @@ class World
         /// Get the maximum skill level a player can reach
         uint16 GetConfigMaxSkillValue() const
         {
-            uint8 lvl = uint8(getIntConfig(CONFIG_MAX_PLAYER_LEVEL));
+            uint16 lvl = uint16(getIntConfig(CONFIG_MAX_PLAYER_LEVEL));
 
             if (lvl >= 1 && lvl < 10)
                 return 75;
